Update RangeDie.java
Shortened casting time: no more double casting from int to float and then to double.

+++ src/main/java/io/github/qwert26/somedice/RangeDie.java
@@ -180,7 +180,7 @@ public final class RangeDie extends AbstractDie {
 	 */
 	@Override
 	public int getDistinctValues() {
-		return (int) Math.ceil((end - start) / (float) step);
+		return (int) Math.ceil((end - start) / (double) step);
 	}
 
 	/**
